feat(example): add shared image marker lab
